IThreadImpl: move platform code to platform dir
