+++ cmake/treedata/freebsd/subdirs.txt
@@ -10,5 +10,6 @@ xbmc/platform/linux/storage         platform/linux/storage
 xbmc/platform/posix                 platform/posix
 xbmc/platform/posix/filesystem      platform/posix/filesystem
 xbmc/platform/posix/network         platform/posix/network
+xbmc/platform/posix/threads         platform/posix/threads
 xbmc/platform/posix/utils           platform/posix/utils
 xbmc/windowing/linux                windowing/linux
