+++ cmake/treedata/osx/subdirs.txt
@@ -12,6 +12,7 @@ xbmc/platform/darwin/utils            platform/darwin/utils
 xbmc/platform/posix                   posix
 xbmc/platform/posix/filesystem        platform/posix/filesystem
 xbmc/platform/posix/network           platform/posix/network
+xbmc/platform/posix/threads           platform/posix/threads
 xbmc/platform/posix/utils             platform/posix/utils
 xbmc/windowing/osx                    windowing/osx
 xbmc/windowing/osx/SDL                windowing/osx/SDL
