+++ cmake/treedata/windowsstore/subdirs.txt
@@ -9,8 +9,8 @@ xbmc/platform/win10/network            platform/win10/network
 xbmc/platform/win10/peripherals        platform/win10/peripherals
 xbmc/platform/win10/powermanagement    platform/win10/powermanagement
 xbmc/platform/win10/storage            platform/win10/storage
+xbmc/platform/win10/threads            platform/win10/threads
 xbmc/platform/win32/filesystem         platform/win32/filesystem
 xbmc/platform/win32/utils              platform/win32/utils
 xbmc/rendering/dx                      rendering/dx
-xbmc/threads/platform/win              threads/win
 xbmc/windowing/win10                   windowing/win10
