+++ xbmc/platform/posix/threads/CMakeLists.txt
@@ -1,0 +1,4 @@
+set(SOURCES ThreadImplPosix.cpp)
+set(HEADERS ThreadImplPosix.h)
+
+core_add_library(platform_posix_threads)
