+++ xbmc/platform/win10/threads/CMakeLists.txt
@@ -1,0 +1,4 @@
+set(SOURCES ${PROJECT_SOURCE_DIR}/xbmc/platform/win32/threads/ThreadImplWin.cpp)
+set(HEADERS ${PROJECT_SOURCE_DIR}/xbmc/platform/win32/threads/ThreadImplWin.h)
+
+core_add_library(threads_win10)
