+++ xbmc/platform/win32/WinMain.cpp
@@ -12,10 +12,10 @@
 #include "platform/Environment.h"
 #include "platform/xbmc.h"
 #include "threads/Thread.h"
-#include "threads/platform/win/Win32Exception.h"
 #include "utils/CharsetConverter.h" // Required to initialize converters before usage
 
 #include "platform/win32/CharsetConverter.h"
+#include "platform/win32/threads/Win32Exception.h"
 
 #include <Objbase.h>
 #include <WinSock2.h>
