+++ xbmc/platform/win32/threads/CMakeLists.txt
@@ -1,0 +1,7 @@
+set(SOURCES ThreadImplWin.cpp
+            Win32Exception.cpp)
+
+set(HEADERS ThreadImplWin.h
+            Win32Exception.h)
+
+core_add_library(threads_win)
