+++ xbmc/platform/win32/threads/ThreadImplWin.h
