+++ xbmc/platform/win32/threads/Win32Exception.cpp
