+++ xbmc/threads/CMakeLists.txt
@@ -18,13 +18,7 @@ set(HEADERS Atomics.h
             IThreadImpl.h)
 
 if(NOT CORE_SYSTEM_NAME STREQUAL windows AND NOT CORE_SYSTEM_NAME STREQUAL windowsstore)
-    list(APPEND SOURCES platform/RecursiveMutex.cpp
-                        platform/pthreads/ThreadImpl.cpp)
-
-    list(APPEND HEADERS platform/pthreads/ThreadImpl.h)
-else()
-    list(APPEND SOURCES platform/win/ThreadImpl.cpp)
-    list(APPEND HEADERS platform/win/ThreadImpl.h)
+    list(APPEND SOURCES platform/RecursiveMutex.cpp)
 endif()
 
 core_add_library(threads)
